requirements.txt
matplotlib==3.10.7
numpy==2.3.4
py_pde==0.47.2
